Prepared move of JdbcValue to mapping.
This is to avoid dependency cycles between mapping and conversion.
