+++ spring-data-jdbc/src/main/java/org/springframework/data/jdbc/core/mapping/JdbcSimpleTypes.java
@@ -31,7 +31,6 @@ import java.util.HashSet;
 import java.util.Set;
 import java.util.UUID;
 
-import org.springframework.data.jdbc.core.convert.JdbcValue;
 import org.springframework.data.mapping.model.SimpleTypeHolder;
 
 /**
