+++ spring-data-jdbc/src/main/java/org/springframework/data/jdbc/core/mapping/JdbcValue.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.springframework.data.jdbc.core.convert;
+package org.springframework.data.jdbc.core.mapping;
 
 import java.sql.JDBCType;
 import java.util.Objects;
@@ -26,17 +26,46 @@ import org.springframework.lang.Nullable;
  * value and the {@link JDBCType} as which a value should get passed to the JDBC driver.
  *
  * @author Jens Schauder
- * @since 1.1
- * @deprecated use {@link org.springframework.data.jdbc.core.mapping.JdbcValue}
+ * @since 2.4
  */
-@Deprecated
-public final class JdbcValue extends org.springframework.data.jdbc.core.mapping.JdbcValue {
+public class JdbcValue {
 
-	private JdbcValue(@Nullable Object value, @Nullable JDBCType jdbcType) {
-		super(value, jdbcType);
+	private final Object value;
+	private final JDBCType jdbcType;
+
+	protected JdbcValue(@Nullable Object value, @Nullable JDBCType jdbcType) {
+
+		this.value = value;
+		this.jdbcType = jdbcType;
 	}
 
 	public static JdbcValue of(@Nullable Object value, @Nullable JDBCType jdbcType) {
 		return new JdbcValue(value, jdbcType);
+	}
+
+	@Nullable
+	public Object getValue() {
+		return this.value;
+	}
+
+	@Nullable
+	public JDBCType getJdbcType() {
+		return this.jdbcType;
+	}
+
+	@Override
+	public boolean equals(Object o) {
+
+		if (this == o)
+			return true;
+		if (o == null || getClass() != o.getClass())
+			return false;
+		JdbcValue jdbcValue = (JdbcValue) o;
+		return Objects.equals(value, jdbcValue.value) && jdbcType == jdbcValue.jdbcType;
+	}
+
+	@Override
+	public int hashCode() {
+		return Objects.hash(value, jdbcType);
 	}
 }
